Add keyboard shortcuts
* Remove markers/arrows using backspace/delete
* Deselect text/marker/arrow using escape
